standard/pyproject.toml
[project]
name = "ai-coding-agent-standard"
version = "0.1.0"
description = "AI Coding Agent - Glass Box Implementation (Standard Edition)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LLM Providers
    "google-genai>=1.0.0",
    "httpx>=0.27.0",  # For Llama (Ollama)
    # Web Fetch
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
    # Config
    "python-dotenv>=1.0.0",  # For .env file support
]

[project.optional-dependencies]
gemini = ["google-genai>=1.0.0"]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
ai-agent-standard = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
